yandex_service/pyproject.toml
[tool.poetry]
name = "service-yandex"
version = "0.1.0"
description = ""
authors = ["saneknapenek <kolesnikovag22@gmail.com>"]
readme = "README.md"
packages = [{include = "service_yandex"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
httpx = "^0.24.1"
uvicorn = "^0.22.0"
celery = {extras = ["redis"], version = "^5.3.1"}
redis-om = "^0.2.1"
sqlalchemy = "^2.0.19"
asyncpg = "^0.28.0"
pillow = "^10.0.0"
imagehash = "^4.3.1"
psycopg2-binary = "^2.9.7"


[tool.poetry.group.dev.dependencies]
environs = "^9.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
